[level 2] Title: 카펫, Time: 0.02 ms, Memory: 10.2 MB -BaekjoonHub

+++ 프로그래머스/lv2/42842. 카펫/카펫.py
@@ -3,7 +3,7 @@ def solution(brown, yellow):
     by = brown + yellow
     
     for h in range(3, by):
-        w = int(by / h)
+        w = by / h
 
         if( ((w * h) == by) & (w >= h) & ( ((w-2) * (h-2)) == yellow)):                
             answer = [w, h]
